Implemented error messages for no auth header

+++ app/recipes/views.py
@@ -2,7 +2,9 @@ from . import recipe_api
 from flask import Blueprint, make_response, request, jsonify
 from app.models.recipe import Recipe
 from app.models.recipeAuth import RecipeApp
+from app.models.category import Category
 from functools import wraps
+import validate
 
 
 def auth(func):
@@ -11,6 +13,9 @@ def auth(func):
     def user_login(*args, **kwargs):
         # Get the access token from the header
         auth_header = request.headers.get('Authorization')
+        if auth_header is None:
+            response = {'message': 'No token provided. Please provide a valid token.'}
+            return make_response(jsonify(response)), 401
         access_token = auth_header.split(" ")[1]
 
         if access_token:
@@ -71,8 +76,13 @@ def create_recipes(user_id, category_id, **kwargs):
                   type: string
                   default: Recipe name already exists.
         """
+    #checks if category exists
+    category = Category.query.filter(Category.user_id==user_id).filter(Category.category_id==category_id).first()
+    if not category:
+        response = {'message': 'Category doesnt exist.'}
+        return make_response(jsonify(response)), 404
     # adds recipes to the database
-    recipe = Recipe.query.filter_by(recipe_name=request.data['recipe_name']).first()
+    recipe = Recipe.query.filter(Recipe.user_id==user_id).filter(Recipe.category_id==category_id).filter_by(recipe_name=request.data['recipe_name']).first()
 
     if not recipe:
         if request.method == "POST":
@@ -80,27 +90,28 @@ def create_recipes(user_id, category_id, **kwargs):
             ingredients = str(request.data.get('ingredients', ''))
             directions = str(request.data.get('directions', ''))
 
+            recipe_name.strip()
+
             if recipe_name:
-                #recipe_name = recipe_name.title()
-                recipe = Recipe(recipe_name=recipe_name, user_id=user_id, category_id=category_id)
-                recipe.save()
-                response = jsonify({
-                    'recipe_id': recipe.recipe_id,
-                    'recipe_name': recipe.recipe_name,
-                    'ingredients': recipe.ingredients,
-                    'directions': recipe.directions,
-                    'date_created': recipe.date_created,
-                    'date_modified': recipe.date_modified,
-                    'created_by' : user_id,
-                    'category_id': recipe.category_id,
-                })
-                return make_response(response), 201
-    else:
-        # There is an existing recipe.
-        response = {
-            'message': 'Recipe already exists.'
-        }
-        return make_response(jsonify(response)), 409
+                if validate.validate_name(recipe_name) == "True":
+                    recipe = Recipe(recipe_name=recipe_name, user_id=user_id, category_id=category_id)
+                    recipe.save()
+                    response = jsonify({
+                        'recipe_id': recipe.recipe_id,
+                        'recipe_name': recipe.recipe_name,
+                        'ingredients': recipe.ingredients,
+                        'directions': recipe.directions,
+                        'date_created': recipe.date_created,
+                        'date_modified': recipe.date_modified,
+                        'created_by' : user_id,
+                        'category_id': recipe.category_id,
+                    })
+                    return make_response(response), 201
+            response = {'message': 'Recipe name required.'}
+            return make_response(jsonify(response)), 400
+    # There is an existing recipe.
+    response = {'message': 'Recipe already exists.'}
+    return make_response(jsonify(response)), 409
 
 
 @recipe_api.route('/categories/<int:category_id>/recipes/', methods=['GET'])
@@ -111,7 +122,6 @@ def view_recipes(user_id, category_id, **kwargs):
         ---
         tags:
           - Recipes
-
         parameters:
           - in: path
             name: category_id
@@ -133,10 +143,8 @@ def view_recipes(user_id, category_id, **kwargs):
             required: false
             type: integer
             description: This route retrieves the specified number of recipes on a page
-
         security:
           - TokenHeader: []
-
         responses:
           200:
             description: You successfully retrieved all recipes
@@ -166,6 +174,11 @@ def view_recipes(user_id, category_id, **kwargs):
                   type: string
                   default: Recipe not found
         """
+    # checks if category exists
+    category = Category.query.filter(Category.user_id==user_id).filter(Category.category_id==category_id)
+    if not category:
+        response = {'message': 'Category name doesnt exist.'}
+        return make_response(jsonify(response)), 422
     # retrieves recipes from the database
     if request.method == "GET":
         page = int(request.args.get('page', 1))
@@ -189,12 +202,10 @@ def view_recipes(user_id, category_id, **kwargs):
                     'category_id': recipe.category_id,
                     }
                 results.append(obj)
-            response = jsonify(results)
-            response.status_code = 200
-            return response
-        return {
-                "message": "There are no recipes in this category or Category doesn't exist"
-                }, 404
+        if results:
+            return make_response(jsonify(results)), 200
+        response = {'message': 'No recipes found'}
+        return make_response(jsonify(response)), 422
 
 
 @recipe_api.route('/categories/<int:category_id>/recipes/<int:recipe_id>', methods=['GET'])
@@ -205,7 +216,6 @@ def view_one_recipe(user_id, category_id, recipe_id, **kwargs):
         ---
         tags:
           - Recipes
-
         parameters:
           - in: path
             name: category_id
@@ -248,13 +258,16 @@ def view_one_recipe(user_id, category_id, recipe_id, **kwargs):
                   type: string
                   default: Recipe not found
         """
+    #checks if category exists
+    category = Category.query.filter(Category.user_id==user_id).filter(Category.category_id==category_id)
+    if not category:
+        response = {'message': 'Category name doesnt exist.'}
+        return make_response(jsonify(response)), 422
     # retrieve a recipe using it's ID
     recipe = Recipe.query.filter_by(recipe_id=recipe_id).first()
     if not recipe:
         # Raise an HTTPException with a 404 not found status code
-        return {
-            "message": "Url doesn't exist. Please type existing url"
-            }, 404
+        return {"message": "No recipe found"}, 404
         # GET
     response = jsonify({
         'recipe_id': recipe.recipe_id,
@@ -329,18 +342,24 @@ def edit_recipe(user_id, category_id, recipe_id, **kwargs):
                   type: string
                   default: No recipe found
         """
+    #checks if category exists
+    category = Category.query.filter(Category.user_id==user_id).filter(Category.category_id==category_id)
+    if not category:
+        response = {'message': 'Category name doesnt exist.'}
+        return make_response(jsonify(response)), 422
     # edit a recipe
     recipe = Recipe.query.filter_by(recipe_id=recipe_id).first()
     if not recipe:
         # Raise an HTTPException with a 404 not found status code
-        return {
-            "message": "Url doesn't exist. Please type existing url"
-            }, 404
+        return {"message": "Url doesn't exist. Please type existing url"}, 404
     if request.method == 'PUT':
         recipe_name = str(request.data.get('recipe_name', ''))
         ingredients = str(request.data.get('ingredients', ''))
         directions = str(request.data.get('directions', ''))
-
+    #
+    # recipe = Recipe.query.filter_by(recipe_name=request.data['recipe_name']).first()
+    # #checks if recipe exists
+    # if recipe != recipe_name:
         recipe.recipe_name = recipe_name
         recipe.ingredients = ingredients
         recipe.directions = directions
@@ -413,15 +432,17 @@ def delete_recipe(user_id, category_id, recipe_id, **kwargs):
                    type: string
                    default: Recipe not found
         """
+    #checks if category exists
+    category = Category.query.filter(Category.user_id==user_id).filter(Category.category_id==category_id)
+    if not category:
+        response = {'message': 'Category name doesnt exist.'}
+        return make_response(jsonify(response)), 422
     # delete a recipe
     recipe = Recipe.query.filter_by(recipe_id=recipe_id).first()
     if not recipe:
         # Raise an HTTPException with a 404 not found status code
-        return {
-            "message": "Url doesn't exist. Please type existing url"
-            }, 404
+        return {"message": "Url doesn't exist. Please type existing url"}, 404
     if request.method == 'DELETE':
         recipe.delete()
-        return {
-            "message": "recipe {} deleted successfully".format(recipe.recipe_id)
+        return {"message": "recipe {} deleted successfully".format(recipe.recipe_id)
             }, 200
